refactor(output): extract server tests to sibling file
